# WIP ARM Neon/AVX2 SIMD implementation.