--- a/ext/json/ext/generator/extconf.rb
+++ b/ext/json/ext/generator/extconf.rb
@@ -8,9 +8,9 @@ else
   $defs << "-DJSON_GENERATOR"
 
   if enable_config('use-simd', default=true)
-    $defs.push("-DENABLE_SIMD")
-
     if RbConfig::CONFIG['host_cpu'] =~ /^(arm.*|aarch64.*)/
+      $defs.push("-DENABLE_SIMD")
+
       # Try to compile a small program using NEON instructions
       if have_header('arm_neon.h')
         have_type('uint8x16_t', headers=['arm_neon.h']) && try_compile(<<~'SRC')
@@ -29,7 +29,30 @@ else
             }
         SRC
         end
-    end
+      elsif have_header('x86intrin.h')
+        
+        # This is currently hardcoded to false as using m256 seems significantly slower on my machine.
+        # TODO make this configurable
+        if false && have_type('__m256i', headers=['x86intrin.h']) && try_compile(<<~'SRC', opt='-mavx2')
+          #include <x86intrin.h>
+          int main() {
+              __m256i test = _mm256_set1_epi8(32);
+              return 0;
+          }
+          SRC
+          $defs.push("-DENABLE_SIMD")
+          append_cflags('-mavx2')
+        elsif have_type('__m128i', headers=['x86intrin.h']) && try_compile(<<~'SRC', opt='-mavx2')
+          #include <x86intrin.h>
+          int main() {
+              __m128i test = _mm_set1_epi8(32);
+              return 0;
+          }
+          SRC
+            $defs.push("-DENABLE_SIMD")
+            append_cflags('-mavx2')
+        end
+      end
   end
 
   create_header