… benchmark/encoder-simple.rb
… ext/json/ext/generator/simd.h
… test/json/json_generator_test.rb