@byroot 2025-01-17
Seems quite a big deal on some benchmarks:

```
== Encoding integers (8009 bytes)
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +YJIT +PRISM [arm64-darwin23]
Warming up --------------------------------------
               after     9.553k i/100ms
Calculating -------------------------------------
               after     96.345k (± 0.5%) i/s   (10.38 μs/i) -    487.203k in   5.056999s

Comparison:
              before:    97143.7 i/s
               after:    96344.5 i/s - same-ish: difference falls within error


== Encoding activitypub.json (52595 bytes)
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +YJIT +PRISM [arm64-darwin23]
Warming up --------------------------------------
               after     2.797k i/100ms
Calculating -------------------------------------
               after     28.500k (± 1.4%) i/s   (35.09 μs/i) -    142.647k in   5.006212s

Comparison:
              before:    21655.6 i/s
               after:    28499.8 i/s - 1.32x  faster


== Encoding citm_catalog.json (500298 bytes)
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +YJIT +PRISM [arm64-darwin23]
Warming up --------------------------------------
               after   134.000 i/100ms
Calculating -------------------------------------
               after      1.356k (± 0.7%) i/s  (737.44 μs/i) -      6.834k in   5.039867s

Comparison:
              before:     1369.0 i/s
               after:     1356.0 i/s - same-ish: difference falls within error


== Encoding twitter.json (466906 bytes)
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +YJIT +PRISM [arm64-darwin23]
Warming up --------------------------------------
               after   277.000 i/100ms
Calculating -------------------------------------
               after      2.773k (± 0.5%) i/s  (360.57 μs/i) -     14.127k in   5.093951s

Comparison:
              before:     2407.9 i/s
               after:     2773.4 i/s - 1.15x  faster
```

I was initially not very enthusiastic about SIMD because we need runtime detection for it to be viable, and that seemed like a lot of work, but I very recently found https://github.com/abetlen/simdinfo, which is simple enough, so it could be worth trying.