@byroot 2025-01-17
> I very recently found https://github.com/abetlen/simdinfo, which is simple enough

Another possibility is https://gcc.gnu.org/wiki/FunctionMultiVersioning, but not too sure what the support is like in various compilers. `ruby/json` being included in Ruby, it's tested against some pretty old and exotic compilers.