@byroot 2025-01-17
> Another possibility is https://gcc.gnu.org/wiki/FunctionMultiVersioning,

Nevermind, it's limited to C++ on i386 arch, so basically unusable.