@samyron 2025-01-18
Apologies about the no `extconf.h`. Forgot to stash that change. I'm currently trying to use the extconf.rb detect NEON (with other implementations coming soon). Additionally trying to add configure flags so we can enable/disable this feature.