@samyron 2025-01-24
Apologies for going dark for a few days. I've been experimenting with a [lookup table based approach](https://github.com/samyron/json/pull/2). I have this implemented in SSE4.2 and have benchmarks included on that PR. I would _expect_ an ARM Neon implementation to be fairly straight forward as Neon provides table lookup instructions directly in their ISA.

The current state of this PR:
1. Implements SIMD Kernels: Arm Neon, AVX2 and SSE4.2 (I need to check the instruction tables again as it might work for SSE2).
2. Determines which kernel to use at runtime for GCC and clang.