@byroot 2025-01-24
Thanks for all your work on this, I'll try to share my thoughts a bit:

I think I'll try to go baby steps. I think at first I'd like to release a version that makes a very small use of simd, as to validate that users don't run into compilation issues, or other runtime issues. Basically validate that the compilation and runtime detection facilities work. (I'm not asking you to simplify your PR, just saying that I will probably end up cherry-picking things from it at first).

Since `ruby/json` is mirrored into Ruby, it will be compiled with tons of different compilers, on lots of weird architectures and OS. So we probably should expect some nasty bug reports.

Then I need to figure out some proper X86 development environment for myself, otherwise it's not really viable. I have a beefy x86 Windows machine, so I can probably run some VM on it.

Once a version is out using some small amount of SIMD and it has been ironed out, I think we'll progressively be able to use it in more places.